Some code cleanup and added unit tests for SearchService.

+++ src/main/java/application/objects/Page.java
@@ -55,7 +55,7 @@ public class Page {
 		outgoingLinks.add(link);
 	}
 
-	public int amountOfOutgoingLinks(){
+	public double amountOfOutgoingLinks(){
 		return outgoingLinks.size();
 	}
 }
